backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
python-dotenv==1.0.0
gunicorn==21.2.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
pydantic==2.1.1
celery==5.3.1
redis==4.6.0
python-multipart==0.0.6
jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
